Refactor AsyncParser and restructure Parser hierarchy.
This commit does a number of things:

  1. Splits the JParser.scala file out into a number of smaller
     files, to make the code a bit more accessible.

  2. Introduce AsyncParse, a type representing the result of an
     asynchronous parse, which contains a Seq[ParseException] and
     a Seq[JValue].

  3. Begin tracking line and column information correctly in Parser,
	 and report this information in ParseException.

  4. Use newlines for error-recovery, according to our previous
     discovery. If the error occurs at a line boundary, the parsing
     will be restarted from that line. Otherwise, the parser will
     scan forward to the next newline and then resume.

  5. JParser.parseAsync has been changed to accept an AsyncParser
     and return (AsyncParse, AsyncParser).

The commit also adds a basic test of the error recovery mechanism,
which ensures that the line/column information for parse errors is
correct.
